requirement.txt
hidapi
tqdm